+++ Sources/NeuralNetwork/Inputs.c
@@ -12,7 +12,7 @@
 #define REWARD_SUPERPACGUM 16
 #define REWARD_FRUIT 100
 
-int *init_inputs()
+double *init_inputs()
 {
   Game *game = get_game();
   int x_blinky = 0;
@@ -34,7 +34,7 @@ int *init_inputs()
   int pos_pacman = X_pc * 28 + Y_pc;
   int map_point = pos_pacman -5 - (5 * 28);
 
-  int *lidar = calloc(11*11,sizeof(int));
+  double *lidar = calloc(11*11,sizeof(int));
   for (size_t i = 0; i < 11; i++)
   {
     for (size_t j = 0; j < 11; j++)
