+++ Sources/NeuralNetwork/Inputs.h
@@ -1,6 +1,6 @@
 #ifndef NEURALNETWORK_INIT_H
 #define NEURALNETWORK_INIT_H
 
-int *init_inputs();
+double *init_inputs();
 
 #endif
