+++ Sources/NeuralNetwork/backpropagation.h
@@ -1,6 +1,6 @@
 #ifndef BACKPROPAGATION
 #define BACKPROPAGATION
 
-void backpropagation(struct Network *network, int *inputs, int output, double derivate_loss);
+void backpropagation(struct Network *network, double *inputs, int output, double derivate_loss);
 
 #endif
